hot-fix(wme-utils): comp. with new descartes client
A change to the AppController's shape discovered in WME v2.290 reports the `descartesClient` has been changed with an additional layer of abstraction, consilidating low-level raw HTTP requests into its own HTTP client named `DescartesHttpClient`. The new `DescartesClient` only stores a reference to it, and its methods are using it to send the HTTP requests, while the new abstraction is intended to prepare the request data.

We however, to resolve entity prototypes had to request features using the controller and mock requests, which creates an abort controller to cancel any previous pending requests to the server. This cauesd race conditions, and if our library has been injected before the WME first loads the map, it would cause the map to not load correctly, because the original abort controller has been trigerred by our mock "request".

To accomodate this issue we decided to patch the controller's abort controller creation method to return a dummy AbortController if we're running the request to make sure we don't interfere with the original implementation. We used to install this patch on `AppController.descartesClient`.

As mentioned, this property has been changed with an additional abstraction layer, and the actual abort controller creation logic has been moved to the `DescartesHttpClient`, a reference to which resides in the `DescartesClient.descartesHttpClient` and resolvable through `AppController.descartesClient`.

To fix the issue while preserving compatability with older WME versions (currently running in prod.), a new utility method has been implemented named `isLegacyDescartesClient` and the patch is installed dynamically based on the descartes client found in runtime.